Extract glob expand to package
